split conformance
