+++ src/helpers/conformance/loadFhirPackageIndex.ts
@@ -1,0 +1,122 @@
+
+import expressions from '../jsonataExpression';
+import { omitKeys } from '../objectFunctions';
+import { isNumeric } from '../stringFunctions';
+import os from 'os';
+import path from 'path';
+import fs from 'fs-extra';
+import { getLogger } from '../logger';
+
+const logger = getLogger();
+let fhirPackageIndex: Record<string, any> = {};
+
+const createPackageIndexFile = (packagePath: string) => {
+  try {
+    const fileList = fs.readdirSync(path.join(packagePath, 'package'));
+    const files = fileList.flatMap((file: string) => {
+      if (file.endsWith('.json') && file !== 'package.json') {
+        // eslint-disable-next-line @typescript-eslint/no-var-requires
+        const content = require(path.join(packagePath, 'package', file));
+        const indexEntry = {
+          filename: file,
+          resourceType: content.resourceType,
+          id: content.id,
+          url: (typeof content.url === 'string' ? content.url : undefined),
+          version: (typeof content.version === 'string' ? content.version : undefined),
+          kind: (typeof content.kind === 'string' ? content.kind : undefined),
+          type: (typeof content.type === 'string' ? content.type : undefined)
+        };
+        return indexEntry;
+      } else {
+        return [];
+      }
+    });
+    fs.writeFileSync(path.join(packagePath, 'package', '.index.json'), JSON.stringify({ 'index-version': 1, files }, null, 2));
+  } catch (e) {
+    logger.error(e);
+    return false;
+  };
+  return true;
+};
+
+const buildFhirCacheIndex = async () => {
+  logger.info('Building global package index (this might take some time...)');
+  const cachePath = path.join(os.homedir(), '.fhir', 'packages');
+  const dirList: string[] = fs.readdirSync(cachePath, { withFileTypes: true }).filter(entry => entry.isDirectory()).map(dir => dir.name);
+  logger.info(`FHIR Packages found in global cache: ${dirList.join(', ')}`);
+  const packageIndexArray = dirList.map(pack => {
+    if (fs.existsSync(path.join(cachePath, pack, 'package', 'package.json'))) {
+      if (!fs.existsSync(path.join(cachePath, pack, 'package', '.index.json'))) {
+        logger.info(`Generating .index.json for package ${pack}...`);
+        createPackageIndexFile(path.join(cachePath, pack));
+      };
+      return {
+        package: pack,
+        path: path.join(cachePath, pack, 'package'),
+        packageManifest: require(path.join(cachePath, pack, 'package', 'package.json')),
+        packageIndex: require(path.join(cachePath, pack, 'package', '.index.json'))
+      };
+    } else {
+      logger.warn(`Folder '${pack}' found in the local FHIR cache does not seem to be a FHIR package (no package.json file found). Skipping it...`);
+      return undefined;
+    };
+  });
+
+  const bindings = {
+    omitKeys,
+    pathJoin: path.join,
+    require: (filePath: string) => {
+      try {
+        // eslint-disable-next-line @typescript-eslint/no-var-requires
+        const payload = require(filePath);
+        return payload;
+      } catch (e) {
+        logger.error(e);
+        throw (e);
+      }
+    },
+    isNumeric
+  };
+  const packageIndexObject = await expressions.createRawPackageIndexObject.evaluate(packageIndexArray, bindings);
+
+  const fixedIndex = await expressions.fixPackageIndexObject.evaluate(packageIndexObject, { isNumeric });
+  return fixedIndex;
+};
+
+export const parseFhirPackageIndex = async () => {
+  const cachePath = path.join(os.homedir(), '.fhir');
+  const fumeIndexPath = path.join(cachePath, 'fume.index.json');
+  if (fs.existsSync(fumeIndexPath)) {
+    logger.info(`Found global package index file at ${fumeIndexPath}`);
+    const dirList: string[] = fs.readdirSync(path.join(cachePath, 'packages'), { withFileTypes: true }).filter(entry => entry.isDirectory()).map(dir => dir.name);
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    const currentIndex = require(fumeIndexPath);
+    const currentPackages = await expressions.extractCurrentPackagesFromIndex.evaluate(currentIndex);
+    const diff: string[] = await expressions.checkPackagesMissingFromIndex.evaluate({ dirList, packages: currentPackages });
+    if (diff.length === 0) {
+      logger.info('Global package index file is up-to-date');
+      return require(fumeIndexPath);
+    } else {
+      logger.info('Global package index file is outdated');
+    };
+  } else {
+    logger.info('Global package index file is missing');
+  }
+  const fumeIndexFile = await buildFhirCacheIndex();
+  fs.writeFileSync(fumeIndexPath, JSON.stringify(fumeIndexFile, null, 2));
+  return fumeIndexFile;
+};
+
+export const getFhirPackageIndex = () => fhirPackageIndex;
+
+export const loadFhirPackageIndex = async () => {
+  try {
+    logger.info('Trying to load global package index...');
+    const index = await parseFhirPackageIndex();
+    fhirPackageIndex = index;
+    return true;
+  } catch (e) {
+    logger.error(e);
+    throw (e);
+  }
+};
