+++ src/helpers/hl7v2/v2normalizeKey.ts
@@ -4,10 +4,10 @@ import { registerV2key } from './registerV2key';
 import { getCache } from '../cache';
 
 export const v2normalizeKey = async (key: string) => {
-  const bindings = { 
-    initCap, 
-    keyMap: getCache().v2keyMap.getDict(), 
-    registerV2key 
+  const bindings = {
+    initCap,
+    keyMap: getCache().v2keyMap.getDict(),
+    registerV2key
   };
   const res = await expressions.v2normalizeKey.evaluate(key, bindings);
   return res;
