+++ src/helpers/runtime/castToFhir.ts
@@ -4,7 +4,7 @@
  */
 
 import thrower from '../thrower';
-import { getStructureDefinition } from '../conformance';
+import conformance from '../conformance';
 import _ from 'lodash';
 import uuidByString from 'uuid-by-string';
 
@@ -28,7 +28,7 @@ const getPrimitiveParser = async (typeName: string): Promise<Function | undefine
   } else {
     // generate and compile the function
     let resFn: Function;
-    const sDef = await getStructureDefinition(typeName);
+    const sDef = await conformance.getStructureDefinition(typeName);
     if (sDef === undefined) return thrower.throwRuntimeError(`error fetching structure definition for type ${typeName}`);
     const valueElementDef = sDef?.snapshot?.element[3]; // 4th element in a primitive's structdef is always the actual primitive value
     // get regular expression string from the standard extension
